Begin Metadata shapes

+++ src/Carbon.Media.Metadata/Options/GainControl.cs
@@ -1,11 +1,11 @@
 ﻿namespace Carbon.Media.Metadata
 {
-    public enum GainControl : short
+    public enum GainControl : byte
     {
-        None = 0,
-        LowGainUp = 1,
-        HighGainUp = 2,
-        LowGainDown = 3,
+        None         = 0,
+        LowGainUp    = 1,
+        HighGainUp   = 2,
+        LowGainDown  = 3,
         HighGainDown = 4
     }
 }
